+++ mirrord/kube/src/api/kubernetes.rs
@@ -16,8 +16,6 @@ use tokio::net::TcpStream;
 use tokio::sync::mpsc;
 use tracing::{info, trace, warn};
 
-#[cfg(feature = "env_guard")]
-use crate::api::env_guard::EnvVarGuard;
 use crate::{
     api::{
         container::{ContainerApi, EphemeralContainer, JobContainer},
@@ -84,7 +82,7 @@ impl AgentManagment for KubernetesAPI {
         .await
         .map_err(|_| KubeApiError::AgentReadyTimeout)??;
 
-        wrap_raw_connection(conn)
+        Ok(wrap_raw_connection(conn))
     }
 
     #[cfg(not(feature = "incluster"))]
@@ -96,7 +94,11 @@ impl AgentManagment for KubernetesAPI {
         trace!("port-forward to pod {}:{}", &pod_agent_name, &agent_port);
         let mut port_forwarder = pod_api.portforward(&pod_agent_name, &[agent_port]).await?;
 
-        wrap_raw_connection(port_forwarder.take_stream(agent_port).unwrap())
+        Ok(wrap_raw_connection(
+            port_forwarder
+                .take_stream(agent_port)
+                .ok_or(KubeApiError::PortForwardFailed)?,
+        ))
     }
 
     async fn create_agent<P>(&self, progress: &P) -> Result<Self::AgentRef, Self::Err>
@@ -145,9 +147,6 @@ impl AgentManagment for KubernetesAPI {
 }
 
 pub async fn create_kube_api(config: Option<LayerConfig>) -> Result<Client> {
-    #[cfg(feature = "env_guard")]
-    let _guard = EnvVarGuard::new();
-
     let mut kube_config = Config::infer().await?;
 
     if let Some(config) = config {
@@ -158,7 +157,6 @@ pub async fn create_kube_api(config: Option<LayerConfig>) -> Result<Client> {
                     .await?;
         }
 
-        #[cfg_attr(not(feature = "env_guard"), allow(unused_mut))]
         if config.accept_invalid_certificates {
             kube_config.accept_invalid_certs = true;
             // Only warn the first time connecting to the agent, not on child processes.
@@ -167,8 +165,6 @@ pub async fn create_kube_api(config: Option<LayerConfig>) -> Result<Client> {
             }
         }
     }
-    #[cfg(feature = "env_guard")]
-    _guard.prepare_config(&mut kube_config);
 
     Client::try_from(kube_config).map_err(KubeApiError::from)
 }