+++ mirrord/layer/src/connection.rs
@@ -1,130 +1,97 @@
 //! Module for the mirrord-layer/mirrord-agent connection mechanism.
 //!
-//! The layer will either start a new agent pod with [`KubernetesAPI`], directly connect to an
-//! existing agent (currently only used for tests), or let the [`OperatorApi`] handle the
-//! connection.
-use std::time::Duration;
+//! The layer will connect to the internal proxy to communicate with the agent.
+use std::net::SocketAddr;
 
-use mirrord_config::LayerConfig;
-use mirrord_kube::{
-    api::{kubernetes::KubernetesAPI, AgentManagment, Connection},
-    error::KubeApiError,
+use futures::{SinkExt, StreamExt};
+use mirrord_protocol::{ClientCodec, ClientMessage, DaemonMessage};
+use tokio::{
+    net::TcpStream,
+    sync::mpsc::{self, Receiver, Sender},
 };
-use mirrord_operator::client::{OperatorApi, OperatorApiError};
-use mirrord_progress::NoProgress;
-use mirrord_protocol::{ClientMessage, DaemonMessage};
-use tokio::sync::mpsc::{Receiver, Sender};
-use tracing::log::info;
+use tracing::log::{error, info};
 
-use crate::{graceful_exit, FAIL_STILL_STUCK};
+use crate::graceful_exit;
 
-/// Helpful message that we print to the user in a [`graceful_exit`] call, when we get a
-/// [`KubeApiError`], or an [`OperatorApiError].
-const FAIL_CREATE_AGENT: &str = r#"
-mirrord-layer failed while trying to establish connection with the agent pod!
+/// Size of the channel used in [`wrap_raw_connection`] for [`ClientMessage`]s.
+const CONNECTION_CHANNEL_SIZE: usize = 1000;
 
-- Suggestions:
+const FAIL_STILL_STUCK: &str = r#"
+- If you're still stuck and everything looks fine:
 
->> Check that the agent pod was created with `$ kubectl get pods`, it should look something like
-   "mirrord-agent-qjys2dg9xk-rgnhr        1/1     Running   0              7s".
+>> Please open a new bug report at https://github.com/metalbear-co/mirrord/issues/new/choose
 
->> Check that you're using the correct kubernetes credentials (and configuration).
+>> Or join our discord https://discord.com/invite/J5YSrStDKD and request help in #mirrord-help.
 
->> Check your kubernetes' context matches where the agent should be spawned.
 "#;
 
-/// Handles a [`KubeApiError`] by printing what went wrong to the user, and closing the  program
-/// with [`graceful_exit`].
-///
-/// Used in [`connect`], as we treat every error there as fatal.
-fn handle_error(err: KubeApiError, config: &LayerConfig) -> ! {
-    match err {
-        KubeApiError::KubeError(kube::Error::HyperError(err)) => {
-            eprintln!("\nmirrord encountered an error accessing the Kubernetes API.\n");
-            if !config.accept_invalid_certificates {
-                eprintln!("Consider passing --accept-invalid-certificates.\n");
-            }
-
-            match err.into_cause() {
-                Some(cause) => graceful_exit!("Exiting due to {}", cause),
-                None => graceful_exit!("mirrord got KubeError::HyperError"),
-            }
-        }
-        _ => graceful_exit!("{FAIL_CREATE_AGENT}{FAIL_STILL_STUCK} with error {err}"),
-    }
-}
-
-/// Calls [`graceful_exit`] when we have an [`OperatorApiError`].
-///
-/// Used in [`connect`] (same as [`handle_error`]), but only when `LayerConfig::operator` is
-/// set to `true` and an operator is installed in the cluster.
-fn handle_operator_error(err: OperatorApiError) -> ! {
-    graceful_exit!("{FAIL_CREATE_AGENT}{FAIL_STILL_STUCK} with error {err}")
-}
-
-/// Initializes the agent pod with [`KubernetesAPI`], and connects to it, returning an API that the
+/// Connects to the internal proxy in given `SocketAddr`
 /// layer uses to communicate with it, in the form of a [`Sender`] for [`ClientMessage`]s, and a
 /// [`Receiver`] for [`DaemonMessage`]s.
-///
-/// ## Direct connection
-///
-/// There is support for connecting directly to an existing pod (instead of going through
-/// `KubernetesAPI::create`) by setting `LayerConfig::connect_tcp`.
-///
-/// This is used for mirrord tests only.
-///
-/// ## Operator
-///
-/// The [`OperatorApi`] takes over the connection procedure if `LayerConfig::operator` is
-/// set to `true` and an operator is installed in the cluster. We use `OperatorApi::discover`
-/// instead of [`KubernetesAPI`].
-pub(crate) async fn connect(
-    config: &LayerConfig,
+pub(crate) async fn connect_to_proxy(
+    addr: SocketAddr,
 ) -> (Sender<ClientMessage>, Receiver<DaemonMessage>) {
-    let progress = NoProgress;
+    let stream = match TcpStream::connect(addr).await {
+        Ok(stream) => stream,
+        Err(e) => {
+            error!("Couldn't connect to internal proxy: {e:?}");
+            graceful_exit!("{FAIL_STILL_STUCK:?}");
+        }
+    };
+    wrap_raw_connection(stream)
+}
 
-    if let Some(address) = &config.connect_tcp {
-        Connection(address)
-            .connect(&progress)
-            .await
-            .unwrap_or_else(|err| handle_error(err, config))
-    } else if config.operator && let Some(connection) = OperatorApi::discover(config).await.transpose() {
-        connection.unwrap_or_else(|err| handle_operator_error(err))
-    } else {
-        let k8s_api = KubernetesAPI::create(config)
-            .await
-            .unwrap_or_else(|err| handle_error(err, config));
+/// Creates the task that handles the messaging between layer/agent.
+/// It does the encoding/decoding of protocol.
+fn wrap_raw_connection(
+    stream: TcpStream,
+) -> (mpsc::Sender<ClientMessage>, mpsc::Receiver<DaemonMessage>) {
+    let mut codec = actix_codec::Framed::new(stream, ClientCodec::new());
 
-        let agent_ref = {
-            if let (Some(pod_agent_name), Some(agent_port)) =
-                (&config.connect_agent_name, config.connect_agent_port)
-            {
-                info!(
-                    "Reusing existing agent {:?}, port {:?}",
-                    pod_agent_name, agent_port
-                );
-                (pod_agent_name.to_owned(), agent_port)
-            } else {
-                let (pod_agent_name, agent_port) = tokio::time::timeout(
-                    Duration::from_secs(config.agent.startup_timeout),
-                    k8s_api.create_agent(&progress),
-                )
-                .await
-                .unwrap_or(Err(KubeApiError::AgentReadyTimeout))
-                .unwrap_or_else(|err| handle_error(err, config));
-                info!("Created new agent {pod_agent_name}, port {agent_port}");
+    let (client_in_tx, mut client_in_rx) = mpsc::channel(CONNECTION_CHANNEL_SIZE);
+    let (daemon_out_tx, daemon_out_rx) = mpsc::channel(CONNECTION_CHANNEL_SIZE);
 
-                // Set env var for children to re-use.
-                std::env::set_var("MIRRORD_CONNECT_AGENT", &pod_agent_name);
-                std::env::set_var("MIRRORD_CONNECT_PORT", agent_port.to_string());
+    tokio::spawn(async move {
+        loop {
+            tokio::select! {
+                msg = client_in_rx.recv() => {
+                    match msg {
+                        Some(msg) => {
+                            if let Err(fail) = codec.send(msg).await {
+                                error!("Error sending client message: {:#?}", fail);
+                                break;
+                            }
+                        }
+                        None => {
+                            info!("mirrord-kube: initiated disconnect from agent");
 
-                (pod_agent_name, agent_port)
+                            break;
+                        }
+                    }
+                }
+                daemon_message = codec.next() => {
+                    match daemon_message {
+                        Some(Ok(msg)) => {
+                            if let Err(fail) = daemon_out_tx.send(msg).await {
+                                error!("DaemonMessage dropped: {:#?}", fail);
+
+                                break;
+                            }
+                        }
+                        Some(Err(err)) => {
+                            error!("Error receiving daemon message: {:?}", err);
+                            break;
+                        }
+                        None => {
+                            error!("agent disconnected");
+
+                            break;
+                        }
+                    }
+                }
             }
-        };
+        }
+    });
 
-        k8s_api
-            .create_connection(agent_ref)
-            .await
-            .unwrap_or_else(|err| handle_error(err, config))
-    }
+    (client_in_tx, daemon_out_rx)
 }