+++ mirrord/layer/tests/sip.rs
@@ -1,6 +1,6 @@
 #![feature(assert_matches)]
 #![cfg(target_os = "macos")]
-use std::{path::PathBuf, time::Duration};
+use std::{path::Path, time::Duration};
 
 use rstest::rstest;
 use tokio::net::TcpListener;
@@ -17,7 +17,7 @@ use mirrord_sip::sip_patch;
 #[rstest]
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
 #[timeout(Duration::from_secs(20))]
-async fn test_tmp_dir_read_locally(dylib_path: &PathBuf) {
+async fn test_tmp_dir_read_locally(dylib_path: &Path) {
     let application = Application::BashShebang;
     let executable = application.get_executable().await;
     let executable = sip_patch(&executable).unwrap().unwrap();
@@ -32,6 +32,8 @@ async fn test_tmp_dir_read_locally(dylib_path: &PathBuf) {
     // Accept the connection from the layer and verify initial messages.
     let mut layer_connection = LayerConnection::get_initialized_connection(&listener).await;
     println!("Application subscribed to port, sending tcp messages.");
+
+    layer_connection.expect_gethostname(1).await;
 
     assert!(layer_connection.is_ended().await);
     test_process.wait().await;